+++ apps/live/src/app/(landing)/CarnivalLanding/Components/OpeningText.tsx
@@ -1,9 +1,9 @@
 "use client";
 
 import React from "react";
-import HBPLogo from "../../../../../main/src/app/lib/Assets/SVG/Hero/LandingAssets/HBPLogo.tsx";
-import CountdownTimer from "./CountdownTimer";
-import UntilSubmissionsClose from "./UntilSubmissionsClose";
+import HBPLogo from "main/src/app/lib/Assets/SVG/Hero/LandingAssets/HBPLogo.tsx";
+import CountdownTimer from "./CountdownTimer.tsx";
+import UntilSubmissionsClose from "../../../lib/Assets/SVG/CarnivalLanding/UntilSubmissionsClose.tsx";
 
 const OpeningText: React.FC = () => {
   return (
