--- apps/live/src/app/(landing)/HeroComponents/LandingGrass.tsx
@@ -1,23 +1,0 @@
-import React from "react";
-const LandingGrass: React.FC = () => {
-  return (
-    <svg
-      width="100%"
-      viewBox="0 0 2000 369"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      style={{ display: "block" }}
-    >
-      <path
-        d="M1438 20.5C1655.5 20.499 2060.5 75.5 2060.5 75.5V225H-60.5V0C-60.5 0 623.6 77.6326 812.5 54.8063C1008.98 31.0646 1220.5 20.501 1438 20.5Z"
-        fill="#4B642E"
-      />
-      <path
-        d="M618.5 80C401 79.999 -60.5 135.5 -60.5 135.5V383.593L2060.5 384V80.4064C2060.5 80.4064 1306.4 157.632 1117.5 134.806C921.024 111.064 836 80.001 618.5 80Z"
-        fill="#709941"
-      />
-    </svg>
-  );
-};
-
-export default LandingGrass;
