+++ apps/live/src/app/lib/Assets/SVG/CarnivalLanding/HeroBackground.tsx
