+++ apps/live/src/app/lib/Assets/SVG/CarnivalLanding/OrangeFirework.tsx
