+++ apps/live/src/app/lib/Assets/SVG/CarnivalLanding/OrangeTent.tsx
