+++ apps/live/src/app/lib/Assets/SVG/CarnivalLanding/PurpleTent.tsx
