+++ apps/live/src/app/lib/Assets/SVG/CarnivalLanding/SubmissionsClose.tsx
