+++ apps/live/src/app/lib/Assets/SVG/CarnivalLanding/UntilSubmissionsClose.tsx
