+++ apps/live/src/app/lib/Assets/SVG/CarnivalLanding/YellowFirework.tsx
